arm: dts: imx93-phyboard-segin-u-boot: Remove unused I2C buses in SPL
On phyCORE-i.MX93 based boards, lpi2c1 is not used in the bootloader nor
later in the Linux. Thus remove it for good.

On the other hand, lpi2c2 is used differently depending on the carrier
board in use (either phyBOARD-Segin-i.MX93 or phyBOARD-Nash-i.MX93). And
since we want bootloader for phyCORE-i.MX93 boards to be carrier board
agnostic, we have no use for lpi2c2 in SPL, thus remove it too.

Signed-off-by: Primoz Fiser <primoz.fiser@norik.com>