board: phytec: phycore-imx93: Use imx93-phyboard-nash device-tree
Use imx93-phyboard-nash for the CONFIG_DEFAULT_DEVICE_TREE instead of
imx93-phyboard-segin due to the fact phyBOARD-Segin is considered as a
legacy board compared to phyBOARD-Nash due to level of compatibility
with other platforms such as i.MX6UL and STM32MP1. As a consequence of
this change, we get two working network interfaces on phyBOARD-Nash in
U-Boot.

Signed-off-by: Primoz Fiser <primoz.fiser@norik.com>